feat(laguna): unify behind test_dflash — single binary, single server
Folds the laguna daemon into test_dflash so server.py spawns one binary for
both qwen35 and laguna targets. Replaces the Phase 3a stopgap (separate
test_laguna_daemon binary + scripts/laguna_serve.py).

New library:
  src/laguna_daemon.h
  src/laguna_daemon.cpp
    dflash27b::run_laguna_daemon(LagunaDaemonArgs)
    Hosts the daemon stdin loop for the Poolside Laguna-XS.2 target on a
    fresh CUDA backend. Speaks both wire formats: server.py'́s bare-prompt
    form (`<prompt_bin> <gen_len>[ samp=...]`, tokens stream as int32 LE on
    --stream-fd terminated by -1) and laguna_pflash_niah.py'́s legacy
    `generate <in> <n_gen> <out>` file-output form. Sampler chain
    (rep_pen → top_k → softmax(temp) → top_p → draw) ported from
    test_dflash.cpp.

test/test_dflash.cpp arch dispatch:
  At main() entry, peek `general.architecture` from the target GGUF via
  gguf_init_from_file. When arch == "laguna":
    - accept the no-draft argv layout (argv[1] = target, argv[2..] = flags)
    - skip qwen35 + DFlash + DDTree init entirely
    - build LagunaDaemonArgs from --max-ctx, --stream-fd, env DFLASH27B_KV_K,
      env DFLASH27B_LAGUNA_CHUNK
    - return run_laguna_daemon(args).
  qwen35 path is byte-identical — no regression risk on existing setups.

test/test_laguna_daemon.cpp now a thin wrapper around run_laguna_daemon.
Kept as a separate binary so scripts/laguna_pflash_niah.py can spawn the
daemon directly (no server.py in the loop) without going through the full
DFlash + DDTree codepath in test_dflash.

scripts/server.py simplification:
  - Always spawn DEFAULT_BIN (test_dflash). Removed DEFAULT_LAGUNA_BIN.
  - For arch=laguna omit --draft + --fast-rollback + --ddtree[-budget],
    keep --daemon + --max-ctx + --stream-fd. test_dflash itself handles
    the rest via its internal arch dispatch.
  - Continues to disable prefix-cache / prefill-cache slots and force PFlash
    compression off when arch=laguna (Phase 2b SNAPSHOT/RESTORE not yet
    landed; sidecar pflash compression not yet wired).

CMakeLists.txt: add src/laguna_daemon.cpp to the dflash27b static library so
both test_dflash and the test_laguna_daemon thin wrapper link against it.

Deleted:
  scripts/laguna_serve.py — superseded by server.py'́s arch-aware spawn.

End-to-end verified on RTX 3090 (Laguna-XS.2 Q4_K_M, max_ctx=1024):
  startup banner    -> [test_dflash] arch=laguna -> dispatching to
                       run_laguna_daemon (max_ctx=1024 kv=q8_0 chunk=2048)
  POST greedy       -> "The capital of France is Paris." (96 tok/s decode)
  POST temp=1.5 s99 -> "Sapphire."  (sampler chain reaches the daemon)

Server.py output now uniform:
  arch      = laguna
  bin       = build/test_dflash   (was: build/test_laguna_daemon)
  draft     = None                (no DFlash draft for laguna yet)
